requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
sqlalchemy>=2.0.0
alembic>=1.13.0
python-jose>=3.3.0
passlib>=1.7.4
python-multipart>=0.0.6
qrcode>=7.4.0
pillow>=10.0.0
python-dotenv>=1.0.0
pydantic-settings>=2.0.0
email-validator>=2.0.0
bcrypt>=4.0.0
aiofiles>=23.0.0
aiohttp>=3.8.0